underWater/projectz_d/requirements.txt
blinker==1.9.0
authlib==1.6.5
certifi==2026.2.25
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
daphne==4.0.0
dotenv==0.9.9
Django==5.2.12
geographiclib==2.1
geopy==2.4.1
gunicorn==22.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
numpy==2.4.2
pandas==3.0.1
pymongo==4.15.3
django-mongodb-backend==5.2.3
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.0
uvloop==0.19.0; python_version < "3.14"
Werkzeug==3.1.6
